feat(features): split implementations into features
